pyproject.toml
[tool.poetry]
name = "mcqgenerator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.33.0"
langchain = "^0.1.16"
openai = "^1.17.1"
colorama = "^0.4.6"
pypdf = "^4.2.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
